requirements.txt
Flask~=2.2.2
timeago~=1.0.16
requests~=2.28.1
img2pdf~=0.4.4
Pillow~=9.2.0
PixivPy3~=3.7.1
pxyTools~=1.1.8